Create HOW TO: Create PHPinfo File and Check PHP Information
The phpinfo() function in PHP is used to display a large amount of information about the current state of PHP. This includes information about PHP compilation options and extensions, the PHP version, server information and environment (if compiled as a module), the PHP environment, OS version information, paths, master and local values of configuration options, HTTP headers, and the PHP license.

Here’s how to use it:

1- Create a PHP File:
2- Create a new PHP file in your hosting document root. Example: phpinfo.php
3- Add the following code in the file created.
4- Open your web browser and navigate to http://your-domain.com/phpinfo.php (replace your-domain.com with your actual domain and phpinfo.php with actual created file name). This will display a page with detailed information about your PHP installation.

Security Note:
It’s important to remove this file after use because it exposes a lot of information about your server's configuration, which could be useful to attackers. To remove the file, simply delete it from your server.